Fix for duplicate index name on the newly partitioned result table. This happens due to running create index statements concurrently, where those create idx statements are creating indexes on a partitioned table.  PG then auto-creates the indexes on the child table, autogenerating names for those indexes.  Since some indexes are in partial create states, PG doesn't seem to fully realize what index names are already in use and ends up creating duplicates. This fix addresses that by 'manually' creating the indexes with unique names on the child tables so PG doesn't have to do it.  When the indexes are created on the parent table, PG should recognize that a compatible index already exists on the child tables and not try to create a new one.
This change requires taking more control over index creation, so the non-configurable index creation classes (which worked via custom db functions) no longer fit the need - they still are in use for the parent result table, but the child tables use a more generic dynamic SQL solution.  This is implemented in the two classes:  CreateBasicIndex and CreateGeomIndex.

The key change is in BuildResultIndexes.  This class now dynamically builds the list of indexes to be created on each child partition table.  One 'win' here is that we should be able to see index creation progress in finer detail as each child index is created.

Other small fixes:
* DateRangePartitionStrategy now uses day of month instead of day of year when creating unique partition table names (improves readability and more reliable name collision avoidance)
* SetupResultTable uses a Bean accessor to fetch ResultPartitionStrategy - that was incorrect.  It now just uses it directly from autowire of a Component.  This caused an issue as soon as I followed the same pattern in the BuildResultIndexes class (duplicate Bean definitions).
* Move a configuration test that was in BuildResultActivityIndexIT to the ConfigurationServiceTest where it belongs.